Populate epic from network
Basic code to fetch epic data over the network.  We aim to keep the
client simple in that we'll use the client side model as the json
string from the server and create the hierarchy of epic-story-task just
in memory for the client.
This code does not handle the network lost case.  We just return false.
 Need to have a last good fetch saved to disk approach.