Add support for network failure and its simulation
We add support for reading back the most recent fetch request when the
network is not available.  We also add a facade to the fetchEpicData
method, so we can also have a simulation of network down supplied, to
exercise the network down code path.

TODO: Need to add a predicate to match the fetch url, instead of just
getting the most recent cached model from the server.