Added code to match cached value to request URL.  Changed test mode back to NoFailureSimulation.
Added a NSPredicate so a matching cache entry to the request we are
making is returned.  In the future we should have the server drive the
tests by having an option in the server to tell the client what test
mode to use, if any, so that we can diagnose apps in the field without
re-deployment.