Allow a demo data project
From reading the human interface guidelines, we note that the user on
first use must be able to assess the quality of the application within
1 or 2 minutes.  This means getting stuck right in with the normal user
flow of adding tasks to an existing project.  It also means not being
burdened by login details or other setup information.  We achieve this
by allowing a new option, "Start Demonstration Project," which
auto-populates the app and takes you to a pre-set epic.  This has
implications for ownership and sharing since no user email address,
project ID, or project name has been set; these have not yet been
worked out.