Add calendar collaboration
The insight here is that tasks are calendar items on a phone because
leverages the schedule ability of a phone.  For example, if you have a
meeting, the tool will place development work around that.  Settings
will allow you to set preferences for when you are working; default
working day is 10am - 6pm 1hr lunch weekdays in your phone's default
timezone.