Updated the server to mirror the client setting the security token
I need to research how GAE does persistence so that the security token
can be remembered.  Also need to change the url tags into the request
enum and then get the string value of them so that we can tightly
encapsulate the protocol into a Request.java file and javadoc comment
it.