Actually persisting new project requests
This change uses Java Persistence API to store on the server New
Project requests from the client.  The client code has been made more
robust to use constants which match the Response.java and Request.java
files.

Did minor fixes to ensure localized strings are tracked or marked not
for localization.

The basic mechanics of server side memory of client side is in place
now.  The next step is to document the user story of trying to create a
new project, listing out the different outcomes, and then implementing
this business logic.