Add the epic view
TODO Need to add json escape code handler both encode and decode to
handle quotes within fields.
TODO Need to add a protocol so we can jump to welcome level and
navigate robot like to the proper project so that create view
controller takes you to the right place.
Updated the wireframe design for the next view, which is the current
projects view.