Add first part of the screen jump protocol.
An ease of use feature, to allow the user to see how the application
screens relate to each other.  When the context indicates a different
screen should be shown, the app will automatically navigate to the
appropriate screen.  We need to put in place the per project screen
first, before finishing off this feature.