Move to a snapshot model
When the UI presents a table, it needs to be stable. That is, not async
updated by contact with the server.  Therefore we have added a means to
snapshot the current model as needed for the current projects UI.  This
also provides a clean-up of details of the model, since the snapshot is
in terms of what is shown on the UI.