Implement automatic screen switching
Using a delay timed invocation, with progmatic highlighting, we now can
animate through from join or create to selection of the project.  We
need to update the async update code to update the UI when the model
updates.  We need a method to poll the server if needed, and call this
each time a data presenting screen is entered.