Remove screen jump protocol
Removing the screen jump protocol.  It looks nice, and was helpful for
the user as it keeps the hierarchy clear.  But it opens us to a lot of
bugs.  The system hints at this with console messages "navigation bar
might be inconsistent" etc.  We've decided to go back to the
traditional use of navigation bar drilling down linearly only.  The
async notifcation from the model back to the ui on the main thread
created the interrupted flow which exposed the problem.